requirements.txt
chromadb==1.4.0
docling==2.66.0
langchain_text_splitters==1.1.0
protobuf==6.33.2
pydantic==2.12.5
pypdf==6.5.0
python-dotenv==1.2.1
Requests==2.32.5
google-genai==0.3.0